+++ examples/arena_vec.rs
@@ -75,7 +75,7 @@ fn main() {
             v.extend_exact([1u32, 2, 3]);
             v.finish()
         };
-        txn.commit();
+        let _ = txn.commit();
         println!("ArenaVec in txn committed");
     }